# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DevJobs",
        "description": "Documentación de la API para la plataforma DevJobs",
        "contact": {
            "name": "David",
            "email": "david@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://devjobs.test",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/vacancies": {
            "get": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Obtener todas las vacantes",
                "operationId": "11d8fef89bfd552fb528f9f9713acaf8",
                "responses": {
                    "200": {
                        "description": "Lista de vacantes obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Vacancy"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Crear una nueva vacante",
                "operationId": "2e10b81243d3bf0bbf5c707d540ceb96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "company",
                                    "last_day",
                                    "salary_id",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Desarrollador PHP"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Se busca desarrollador con experiencia en Laravel."
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "TechCorp"
                                    },
                                    "last_day": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-30"
                                    },
                                    "salary_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vacante creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Vacancy"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para crear vacantes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vacancies/{id}": {
            "get": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Obtener una vacante por ID",
                "operationId": "d250c09acd5ea6f61d6a254351cfbf43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la vacante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vacante encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Vacancy"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vacante no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Actualizar una vacante existente",
                "operationId": "b4185cfe77fd186b7ce95e2bd430da7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la vacante a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Frontend Developer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Experiencia con Vue.js o React."
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "WebDev Inc."
                                    },
                                    "last_day": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-15"
                                    },
                                    "salary_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacante actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Vacancy"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para actualizar esta vacante"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Eliminar una vacante",
                "operationId": "854b113b965ce09b72f929b71a9cb734",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la vacante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vacante eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vacancy deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para eliminar esta vacante"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-vacancies": {
            "get": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Obtener las vacantes propias según el rol",
                "operationId": "0e3eba73230009e5e2e83f35731ec29e",
                "responses": {
                    "200": {
                        "description": "Lista de vacantes filtradas por el rol del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Vacancy"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 15
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para ver vacantes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Vacancy": {
                "title": "Vacancy",
                "required": [
                    "id",
                    "title",
                    "description",
                    "company",
                    "last_day",
                    "salary_id",
                    "category_id",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Desarrollador Laravel"
                    },
                    "description": {
                        "type": "string",
                        "example": "Se busca programador con experiencia en Laravel."
                    },
                    "company": {
                        "type": "string",
                        "example": "Tech Solutions"
                    },
                    "last_day": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-07-01"
                    },
                    "salary_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-06T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-06T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}